[#1474] Fix barrier in mac os but restricting the count to u16

+++ iceoryx2-bb/concurrency/src/internal/strategy/barrier.rs
@@ -15,7 +15,7 @@
 pub struct Barrier(internal::Barrier);
 
 impl Barrier {
-    pub fn new(n: u32) -> Self {
+    pub fn new(n: u16) -> Self {
         Self(internal::Barrier::new(n))
     }
 }
